refactor(domains): issueLottos, WinningStatistics 파일 common 폴더로 이동
- web 폴더와 같은 depth에 두기 위함
- 파일 역할에 맞춰 폴더를 명확히 분리하기 위함
