+++ src/domains/common/issueLottos.js
@@ -1,10 +1,10 @@
-import getUniqueRandomNumbers from "../utils/getUniqueRandomNumbers.js";
+import getUniqueRandomNumbers from "../../utils/getUniqueRandomNumbers.js";
 import {
   MIN_UNIT,
   MIN_LOTTO_NUMBER,
   MAX_LOTTO_NUMBER,
   LOTTO_LENGTH,
-} from "../constants/constants.js";
+} from "../../constants/constants.js";
 
 const issueLottos = (purchaseAmount) => {
   const lottoCount = purchaseAmount / MIN_UNIT;
